corrigindo css

+++ chef-em-casa-front/src/App.js
@@ -1,5 +1,5 @@
 
-import React from "react";
+import React, { useEffect } from "react";
 import './App.css';
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
@@ -9,7 +9,14 @@ import Recipe from "./pages/Recipe";
 import PrivateRoute from "./Routes/PrivateRoute";
 
 function App() {
-  localStorage.clear();
+  useEffect(() => {
+    // Verifica se a aplicação já foi inicializada
+    if (!localStorage.getItem("initialized")) {
+      localStorage.clear(); // Limpa todos os dados
+      localStorage.setItem("initialized", "true"); // Marca como inicializado
+    }
+  }, []);
+
   return (
     <Router>
       <Routes>
